Move some parts of the integration tests into a separate file.
