+++ Cargo.toml
@@ -27,6 +27,10 @@ num-traits = "0.2.14"
 tracing = "0.1.37"
 uuid = "0.8.2"
 
+[[test]]
+name = "integration"
+path = "tests/integration.rs"
+
 [dependencies.clap]
 version = "4"
 default-features = false
